current file

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,25 +1,26 @@
 #include "lists.h"
+
 /**
- * 
- *
- *
+ * print_list - print the list
+ * @h: pointer to structur
+ * Return: return value size
  */
 
 size_t print_list(const list_t *h)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	if (!h)
-        {
-                return (0);
-        }
-        while (h)
+	{
+		return (0);
+	}
+	while (h)
 	{
 		if (h->str)
 			printf("[%u], %s\n", h->len, h->str);
 		else
 			printf("[0] (nil)\n");
-		h = h -> next;
+		h = h->next;
 		i++;
 	}
 	return (i);
